separate stream set and stream manager
